Let's allow code blocks to break across pages.
It would probably be ideal to allow a per-project default to be set, with inline asciidoc overrides. Barring that, it seems best to default to the standard used by the Prags and others. Most people can deal with a code break.

Also, for books with code listings, forcing code to stay together causes way too much headache.